requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
future==0.16.0
giphy-client==1.0.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-twitter==3.4.2
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
